requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.2
aiohttp==3.9.3
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
